+++ NEWS.md
@@ -1,5 +1,7 @@
 # r2dii.match (development version)
 
+* `match_name` now accepts loanbook input with `sector` and `borderline` 
+  already defined (#326). 
 * Avoid failures due to changes in r2dii.data that are expected.
 
 # r2dii.match 0.0.6